source-codes/batch-job/requirements.txt
google-cloud-batch
google-api-python-client
google-cloud-logging